+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed documentation and validation checks for `FBetaMultiLabelMetric`.
 - Fixed handling of HTTP errors when fetching remote resources with `cached_path()`. Previously the content would be cached even when
   certain errors - like 404s - occurred. Now an `HTTPError` will be raised whenever the HTTP response is not OK.
+- Ensured that `MeanAbsoluteError` always returns a `float` metric value instead of a `Tensor`.
 
 
 ## [v1.4.1](https://github.com/allenai/allennlp/releases/tag/v1.4.1) - 2021-01-29